Try out generalising vectors using "monoids"
And using context and so on. But in the end, I think the API I've developed here is not very nice and intuitive. Especially because you'd have to write `with(IntMonoid) {` around reusable parts of the code. Really not happy with the amount of extra code I'd have to write in the part where I use the API then :(